chore: move shared styles to layout; fix mobile light /dark mode issues; make account route catch all
